agent-service/requirements.txt
livekit-agents>=1.4.0
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
aiohttp
